% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{as.2DSample}
\alias{as.2DSample}
\alias{as.2DSample1}
\alias{as.2DSample2}
\alias{as.2DSample3}
\title{Transform data.frame into an object of type 2DSample}
\usage{
as.2DSample(df, dim1, dim2)
}
\arguments{
\item{df}{a dataframe}

\item{dim1}{the first column index as a number or name
as a string}

\item{dim2}{the second column index as a number or name
as a string}
}
\value{
A list with two vectors of x and y of the
samples coordinates. The two list members are
also named x and y and the list itself is of
class type \code{2DSample}.
}
\description{
\code{as.2DSample} takes a dataframe and extracts 2
columns of it and represent those as a two-dimensional
joint sample.
}
\concept{Methods for 2dSample class}
