% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{as.joint_sample}
\alias{as.joint_sample}
\alias{as.joint_sample1}
\alias{as.joint_sample2}
\alias{as.joint_sample3}
\title{Transform data.frame into an object of type jointSample}
\usage{
as.joint_sample(df, dim1, dim2)
}
\arguments{
\item{df}{a dataframe}

\item{dim1}{the first column}

\item{dim2}{the second column}
}
\value{
A list with two vectors of x and y of the samples
coordinates. The two list members are also named
x and y and the list itself is of class type
\code{2dJointSample}.
}
\description{
\code{as.joint_sample} takes a dataframe and extracts 2
columns of it and represent those as a two-dimensional
joint sample.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\seealso{
\code{\link{frown}},\code{\link{smile}},
\code{\link{circle}}, \code{\link{pizza}}, \code{\link{square}}.

Other Two-dimensionalsampling functions.: 
\code{\link{Gaussian}()},
\code{\link{circle}()},
\code{\link{copMix}()},
\code{\link{pizza}()},
\code{\link{rMIErlang}()},
\code{\link{rMIE}()},
\code{\link{square}()},
\code{\link{triangle}()}
}
\concept{Two-dimensionalsampling functions.}
