% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{circle}
\alias{circle}
\alias{circle1}
\alias{circle2}
\alias{circle3}
\title{Simulate a Uniform distribution on a circle}
\usage{
circle(n, r = 1, x0 = 0, y0 = 0, from = 0, to = 1)
}
\arguments{
\item{n}{the size of the sample}

\item{r}{the radius of the circle, default to \code{1}}

\item{x0}{x coordinate of the center of the head, default to \code{0}}

\item{y0}{y coordinate of the center of the head, default to \code{0}}

\item{from}{the radius of the mouth, default to \code{0.05}}

\item{to}{the radius of the mouth, default to \code{0.45}}
}
\value{
A list with two vectors of x and y of the samples coordinates.
The two list members are also named x and y and the list
itself is of class type \code{2dJointSample}.
}
\description{
\code{circle} simulates points in the plan as if they
were distributed uniformly on a circle.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.

Most functions have three tags: @param, @examples and @return.
}
\examples{
mySample <- circle(5000)

}
\seealso{
\url{https://www.r-project.org}

\code{\link{frown}} for products,
\code{\link{cumsum}} for cumulative sums, and
\code{\link{colSums}}/\code{\link{rowSums}}
marginal sums over high-dimensional arrays.
}
\concept{Two-dimensionalsampling functions.}
