% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{common_scale}
\alias{common_scale}
\alias{common_scale1}
\alias{common_scale2}
\alias{common_scale3}
\title{Generate the empirical copula from a joint random sample}
\usage{
common_scale(
  n,
  scale_model = Exponential(100),
  margin_models = list(x = Exponential(100), y = Exponential(100))
)
}
\arguments{
\item{n}{an object of type \code{2DSample}}

\item{scale_model}{an object of type \code{2DSample}}

\item{margin_models}{an object of type \code{2DSample}}
}
\value{
\preformatted{    The two list members are also named x and y and the list
    itself is of class type \code{2DSample}.
}

The empirical copula model of the joint sampleas a list
with two vectors of x and y of the samples coordinates.
}
\description{
\code{copula} generates the empirical copula
corresponding to the joint sample passed in
as input
}
\examples{
common_scale(frown(1000))

}
