% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtures.R
\name{copMix}
\alias{copMix}
\alias{copMix1}
\alias{copMix2}
\alias{copMix3}
\title{Mixture of copulas}
\usage{
copMix(copulas, weights)
}
\arguments{
\item{copulas}{list of copulas}

\item{weights}{the vector of weights}
}
\value{
A list with two vectors of x and y of the samples
coordinates. The two list members are also named
x and y and the list itself is of class type
\code{2dJointSample}.
}
\description{
\code{copMix} takes n empirical copulas and as many
wwights produces a mixture of copulas
}
\examples{
Gaussian(5000, c(3,4), matrix(c(25, 8, 8, 36), 2) )

}
\seealso{
Other Two-dimensionalsampling functions.: 
\code{\link{rMIErlang}()},
\code{\link{rMIE}()}
}
\concept{Two-dimensionalsampling functions.}
