% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVX.R
\docType{package}
\name{CVX}
\alias{CVX}
\title{CVX: A package for computating the notorious bar statistic}
\description{
The CVX package provides three categories of important functions:
foo, bar and baz.
}
\section{Simple two dimensional joint sample models}{

\itemize{
\item \link{frown}
\item \link{smile}
\item \link{circle}
\item \link{pizza}
\item \link{square}
\item \link{triangle}
}
}

\section{Copula models}{


\code{CommonShock}
\code{ErlangMixtures}
}

\section{Utility functions}{


The foo functions ...
}

\section{Other functions}{

The foo functions ...
}

