% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Degenerate}
\alias{Degenerate}
\alias{Degenerate1}
\alias{Degenerate2}
\alias{Degenerate3}
\title{Degenerate Distribution with parameter}
\usage{
Degenerate(params)
}
\arguments{
\item{params}{the size of the sample}
}
\value{
a list with two vectors of x and y of the samples coordinates.
The two list members are also named x and y and the list
itself is of class type \code{2DSample}.
}
\description{
\code{frown} simulates points in the plan
as if they were distributed uniformly on
a drawing of a frowning face.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\examples{
Degenerate(c(1000))

}
\seealso{
\code{\link{Uniform}},
\code{\link{Pareto}}, \code{\link{InversePareto}},
\code{\link{Exponential}}.

Other Two-dimensional sampling functions.: 
\code{\link{Exponential}()},
\code{\link{InversePareto}()},
\code{\link{Pareto}()}
}
\concept{Two-dimensional sampling functions.}
