% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Exponential}
\alias{Exponential}
\alias{Exponential1}
\alias{Exponential2}
\alias{Exponential3}
\title{Simulate a Uniform distribution on a frowning face}
\usage{
Exponential(params)
}
\arguments{
\item{params}{the size of the sample}
}
\value{
a list with two vectors of x and y of the samples coordinates.
The two list members are also named x and y and the list
itself is of class type \code{2dJointSample}.
}
\description{
\code{frown} simulates points in the plan
as if they were distributed uniformly on
a drawing of a frowning face.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\seealso{
\code{\link{triangle}},\code{\link{smile}},
\code{\link{circle}}, \code{\link{pizza}}, \code{\link{square}}.

Other Two-dimensional sampling functions.: 
\code{\link{frown}()},
\code{\link{plot.jointSample}()},
\code{\link{print.jointSample}()},
\code{\link{smile}()}
}
\concept{Two-dimensional sampling functions.}
