% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{frown}
\alias{frown}
\alias{frown1}
\alias{frown2}
\alias{frown3}
\title{Draws points uniformly distributed on a frowning face}
\usage{
frown(
  n,
  x0 = 0,
  y0 = 0,
  main_radius = 1,
  eye_radius = 0.25,
  mouth_radius = 0.7,
  mouth_start = 0.05,
  mouth_end = 0.45
)

frown(
  n,
  x0 = 0,
  y0 = 0,
  main_radius = 1,
  eye_radius = 0.25,
  mouth_radius = 0.7,
  mouth_start = 0.05,
  mouth_end = 0.45
)
}
\arguments{
\item{n}{the size of the sample}

\item{x0}{x coordinate of the center of the head,
default to \code{0}}

\item{y0}{y coordinate of the center of the head,
default to \code{0}}

\item{main_radius}{the radius of the head, default
to \code{1}}

\item{eye_radius}{the radius of the eye, default
to \code{0.25}}

\item{mouth_radius}{the radius of the mouth,
default to \code{0.7}}

\item{mouth_start}{number between \code{0} and \code{1}
representing the fraction of \code{(2xPi)} angle
at which the mouth stroke begins, default to
\code{0.05}}

\item{mouth_end}{number between \code{0} and \code{1}
representing the fraction of \code{(2Pi)} angle
at which the mouth stroke ends, default to
\code{0.45}}
}
\value{
a list with two vectors of x and y of the
samples coordinates. The two list members
are also named x and y and the list itself
is of class type \code{2DSample}.

a list with two vectors of x and y of the
samples coordinates. The two list members
are also named x and y and the list itself
is of class type \code{2DSample}.
}
\description{
\code{frown} simulates points in the plan
as if they were distributed uniformly on
the drawing of a frowning face.

\code{frown} simulates points in the plan
as if they were distributed uniformly on
the drawing of a frowning face.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.

This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\examples{
frown(5000)

frown(5000)

}
\seealso{
Other Two-dimensional sampling functions: 
\code{\link{circle}()},
\code{\link{pizza}()},
\code{\link{smile}()},
\code{\link{square}()},
\code{\link{triangle}()}

Other Two-dimensional sampling functions: 
\code{\link{circle}()},
\code{\link{pizza}()},
\code{\link{smile}()},
\code{\link{square}()},
\code{\link{triangle}()}
}
\concept{Two-dimensional sampling functions}
