% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{InversePareto}
\alias{InversePareto}
\alias{InversePareto1}
\alias{InversePareto2}
\alias{InversePareto3}
\title{Inverse Pareto Distribution}
\usage{
InversePareto(params)
}
\arguments{
\item{params}{the size of the sample}
}
\value{
a list with two vectors of x and y of the samples coordinates.
The two list members are also named x and y and the list
itself is of class type \code{2dJointSample}.
}
\description{
\code{frown} simulates points in the plan
as if they were distributed uniformly on
a drawing of a frowning face.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\examples{
InversePareto(c(1000, 3))

}
\seealso{
\code{\link{Pareto}},\code{\link{Exponential}},
\code{\link{Uniform}}, \code{\link{Degenerate}},

Other Two-dimensional sampling functions.: 
\code{\link{Degenerate}()},
\code{\link{Exponential}()},
\code{\link{Pareto}()},
\code{\link{frown}()},
\code{\link{plot.jointSample}()},
\code{\link{print.jointSample}()},
\code{\link{smile}()}
}
\concept{Two-dimensional sampling functions.}
