% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{Movie}
\alias{Movie}
\alias{Movie1}
\alias{Movie2}
\alias{Movie3}
\title{Make a movie from going from one plot to another one
Generate the empirical copula from a joint random sample}
\usage{
Movie(start_point, end_point, steps)
}
\arguments{
\item{start_point}{an object of type \code{2DSample}}

\item{end_point}{an object of type \code{2DSample}}

\item{steps}{an object of type \code{2DSample}}
}
\value{
\preformatted{    The two list members are also named x and y and the list
    itself is of class type \code{2DSample}.
}

The empirical copula model of the joint sampleas a list
with two vectors of x and y of the samples coordinates.
}
\description{
\code{copula} generates the empirical copula
corresponding to the joint sample passed in
as input
}
\examples{
fr <- frown(1000)
Movie(fr, copula(fr), 100)

}
