% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{pizza}
\alias{pizza}
\alias{pizza1}
\alias{pizza2}
\alias{pizza3}
\title{Simulate a Uniform distribution on a pizza slice}
\usage{
pizza(n, r = 1, hole = 0, x0 = 0, y0 = 0, from = 0, to = 1)
}
\arguments{
\item{n}{the size of the sample}

\item{r}{the radius of the disk, default to \code{1}}

\item{hole}{the radius of the hole in the center of the
pizza, default to \code{0}}

\item{x0}{x coordinate of the center of the disk, default to \code{0}}

\item{y0}{y coordinate of the center of the disk, default to \code{0}}

\item{from}{starting point of the arc as a fraction of the whole
circumference starting from the rightmost point and
drawing the circle in a counter-clockwise direction,
default to \code{0}}

\item{to}{ending point of the arc as a fraction of the whole
circumference starting from the rightmost point and
drawing the circle in a counter-clockwise direction,
default to \code{0}}
}
\value{
A list with two vectors of x and y of the samples coordinates.
The two list members are also named x and y and the list
itself is of class type \code{2dJointSample}.
}
\description{
\code{pizza} simulates points in the plan as if they
were distributed uniformly inside a pizza slice.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\examples{
pizza(5000)

}
\seealso{
\code{\link{frown}},\code{\link{smile}},
\code{\link{circle}}, \code{\link{pizza}}, \code{\link{square}}.

Other Two-dimensionalsampling functions.: 
\code{\link{circle}()},
\code{\link{square}()},
\code{\link{triangle}()}
}
\concept{Two-dimensionalsampling functions.}
