% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{plot.2DJointSample}
\alias{plot.2DJointSample}
\alias{plot.2DJointSample1}
\alias{plot.2DJointSample2}
\alias{plot.2DJointSample3}
\title{Plot a joint sample}
\usage{
\method{plot}{`2DJointSample`}(x, pch = 20, cex = 0.1, all = F, ...)
}
\arguments{
\item{x}{the size of the sample}

\item{pch}{the size of the sample}

\item{cex}{the size of the sample}

\item{all}{the size of the sample}

\item{...}{the size of the sample}
}
\value{
a list with two vectors of x and y of the
samples coordinates. The two list members are
also named x and y and the list itself is of
class type \code{2dJointSample}.
}
\description{
\code{plot.2DJointSample} simulates points in the plan
as if they were distributed uniformly on
a drawing of a frowning face.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\examples{
plot(frown(5000))

}
\seealso{
Other Methods for 2dJointSample class: 
\code{\link{print.2DJointSample}()}
}
\concept{Methods for 2dJointSample class}
