% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{plot.jointSample}
\alias{plot.jointSample}
\alias{plot.jointSample1}
\alias{plot.jointSample2}
\alias{plot.jointSample3}
\title{Plot a joint sample}
\usage{
\method{plot}{jointSample}(x, pch = 20, cex = 0.1, all = F, ...)
}
\arguments{
\item{x}{the size of the sample}

\item{pch}{the size of the sample}

\item{cex}{the size of the sample}

\item{all}{the size of the sample}

\item{...}{the size of the sample}
}
\value{
a list with two vectors of x and y of the samples coordinates.
The two list members are also named x and y and the list
itself is of class type \code{2dJointSample}.
}
\description{
\code{frown} simulates points in the plan
as if they were distributed uniformly on
a drawing of a frowning face.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\examples{
plot(frown(5000), 15, 0.2, F, ...)

}
\seealso{
\code{\link{triangle}},\code{\link{smile}},
\code{\link{circle}}, \code{\link{pizza}}, \code{\link{square}}.

Other Two-dimensional sampling functions.: 
\code{\link{Exponential}()},
\code{\link{frown}()},
\code{\link{print.jointSample}()},
\code{\link{smile}()}
}
\concept{Two-dimensional sampling functions.}
