% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{print.2DJointSample}
\alias{print.2DJointSample}
\alias{print.jointSample1}
\alias{print.jointSample2}
\alias{print.jointSample3}
\title{Print method for the jointSample type}
\usage{
\method{print}{`2DJointSample`}(x, ...)
}
\arguments{
\item{x}{the size of the sample}

\item{...}{the size of the sample}
}
\value{
a list with two vectors of x and y of the samples coordinates.
The two list members are also named x and y and the list
itself is of class type \code{2dJointSample}.
}
\description{
\code{print.2DJointSample} simulates points in the plan
as if they were distributed uniformly on
a drawing of a frowning face.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\examples{
print(frown(5000))

}
\seealso{
Other Methods for 2dJointSample class: 
\code{\link{plot.2DJointSample}()}
}
\concept{Methods for 2dJointSample class}
