% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{print.2DSample}
\alias{print.2DSample}
\alias{print.2DSample1}
\alias{print.2DSample2}
\alias{print.2DSample3}
\title{Print a two-dimensional joint sample}
\usage{
\method{print}{`2DSample`}(x, ...)
}
\arguments{
\item{x}{the sample itself}

\item{...}{optional parameters}
}
\description{
\code{print.2DSample} print method for objects of
class \code{2DSample}.
}
\examples{
print(frown(5000))

}
\seealso{
Other Methods for 2DSample class: 
\code{\link{plot.2DSample}()}
}
\concept{Methods for 2DSample class}
