% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{print.jointSample}
\alias{print.jointSample}
\alias{print.jointSample1}
\alias{print.jointSample2}
\alias{print.jointSample3}
\title{Simulate a Uniform distribution on a frowning face}
\usage{
\method{print}{jointSample}(x, ...)
}
\arguments{
\item{x}{the size of the sample}

\item{...}{the size of the sample}
}
\value{
a list with two vectors of x and y of the samples coordinates.
The two list members are also named x and y and the list
itself is of class type \code{2dJointSample}.
}
\description{
\code{frown} simulates points in the plan
as if they were distributed uniformly on
a drawing of a frowning face.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\examples{
print(frown(5000))

}
\seealso{
\code{\link{triangle}},\code{\link{smile}},
\code{\link{circle}}, \code{\link{pizza}}, \code{\link{square}}.

Other Two-dimensional sampling functions.: 
\code{\link{Degenerate}()},
\code{\link{Exponential}()},
\code{\link{InversePareto}()},
\code{\link{Pareto}()},
\code{\link{frown}()},
\code{\link{plot.jointSample}()},
\code{\link{smile}()}
}
\concept{Two-dimensional sampling functions.}
