% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{rand_copula}
\alias{rand_copula}
\alias{rand_copula1}
\alias{rand_copula2}
\alias{rand_copula3}
\title{Apply a different dependency structure to a joint
random sample.}
\usage{
rand_copula(cop, sample)
}
\arguments{
\item{cop}{an empirical copua object}

\item{sample}{an object of type \code{2dJointSample}}
}
\value{
\preformatted{    The two list members are also named x and y and the list
    itself is of class type \code{2dJointSample}.
}

A jointsample with thesame copula as \code{cop} and
the same marginals as \code{sample}
}
\description{
\code{rand_copula} generates a joint random sample
with the same marginals as the sample passed in as
and the empirical copula also passed in.
}
\examples{
fr <- frown(1000)
sm <- smile(1000)
rand_copula(copula(fr), sm)

}
