% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{randomize}
\alias{randomize}
\alias{randomize1}
\alias{randomize2}
\alias{randomize3}
\title{Generate the empirical copula from a joint random sample}
\usage{
randomize(x)
}
\arguments{
\item{x}{vector for which values are randomized}
}
\value{
\preformatted{    The two list members are also named x and y and the list
    itself is of class type \code{2DSample}.
}
}
\description{
\code{copula} generates the empirical copula
corresponding to the joint sample passed in
as input
}
\examples{
randomize(runif(100))

}
