% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erlang.R
\name{rErlang}
\alias{rErlang}
\title{\code{rErlang} simulates points in the plan as if they
were distributed as a mixture of Independent Erlang
distributions with fixed scale parameter.}
\usage{
rErlang(sample_size, scale, shape)
}
\arguments{
\item{sample_size}{size ofsample}

\item{scale}{commonscale parameter}

\item{shape}{vector of shape parameters}
}
\value{
If l = length(shapes), then the function
returns a matrix of sample_size rows and
l columns.
}
\description{
\code{rErlang} simulates points in the plan as if they
were distributed as a mixture of Independent Erlang
distributions with fixed scale parameter.
}
