% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erlang.R
\name{rIErlang}
\alias{rIErlang}
\title{Random realization vectors of Independent Erlangs}
\usage{
rIErlang(sample_size, scale, shapes, weights)
}
\arguments{
\item{sample_size}{size ofsample}

\item{scale}{commonscale parameter}

\item{shapes}{vector of shape parameters}

\item{weights}{vector of mixture weights}
}
\value{
If l = length(shapes), then the function
returns a matrix of sample_size rows and
l columns.
}
\description{
\code{rIErlang} simulates points in the plan as if they
were distributed as a mixture of Independent Erlang
distributions with fixed scale parameter.
}
