% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erlang.R
\name{rMErlang}
\alias{rMErlang}
\title{Random realization vectors of Independent Erlangs}
\usage{
rMErlang(sample_size, scale, weights, shapes)
}
\arguments{
\item{sample_size}{size ofsample}

\item{scale}{commonscale parameter}

\item{weights}{vector of mixture weights}

\item{shapes}{vector of shape parameters}
}
\value{
If l = length(shapes), then the function
returns a matrix of sample_size rows and
l columns.
}
\description{
\code{rMIErlang} simulates points in the plan as if they
were distributed as a mixture of Independent Erlang
distributions with fixed scale parameter.
}
