% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtures.R
\name{rMIE}
\alias{rMIE}
\alias{rMIE1}
\alias{rMIE2}
\alias{rMIE3}
\title{Mixture of Independent Erlangs#'
\code{copMix} takes n empirical copulas and as many
wwights produces a mixture of copulas}
\usage{
rMIE(sample_size, dimension, scale, weights, shapes)
}
\arguments{
\item{sample_size}{list of copulas}

\item{dimension}{the vector of weights}

\item{scale}{list of copulas}

\item{weights}{the vector of weights}

\item{shapes}{list of copulas}
}
\value{
A list with two vectors of x and y of the samples
coordinates. The two list members are also named
x and y and the list itself is of class type
\code{2dJointSample}.
}
\description{
Mixture of Independent Erlangs#'
\code{copMix} takes n empirical copulas and as many
wwights produces a mixture of copulas
}
\seealso{
Other Two-dimensionalsampling functions.: 
\code{\link{Gaussian}()},
\code{\link{as.joint_sample}()},
\code{\link{circle}()},
\code{\link{copMix}()},
\code{\link{pizza}()},
\code{\link{rMIErlang}()},
\code{\link{square}()},
\code{\link{triangle}()}
}
\concept{Two-dimensionalsampling functions.}
