% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erlang.R
\name{rMIErlang}
\alias{rMIErlang}
\alias{rMIErlang1}
\alias{rMIErlang2}
\alias{rMIErlang3}
\title{Simulate a mixture of Independent Erlang distribution}
\usage{
rMIErlang(sample_size, scale, weights, shapes)
}
\arguments{
\item{sample_size}{size ofsample}

\item{scale}{commonscale parameter}

\item{weights}{vector of mixture weights}

\item{shapes}{vector of shape parameters}
}
\value{
A list with two vectors of x and y of the samples
coordinates. The two list members are also named
x and y and the list itself is of class type
\code{2DSample}.
}
\description{
\code{rMIErlang} simulates points in the plan as if they
were distributed as a mixture of Independent Erlang
distributions with fixed scale parameter.
}
\seealso{
Other Two-dimensionalsampling functions.: 
\code{\link{copMix}()},
\code{\link{rMIE}()}
}
\concept{Two-dimensionalsampling functions.}
