% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{smile}
\alias{smile}
\alias{smile1}
\alias{smile2}
\alias{smile3}
\title{Simulate a Uniform distribution on a smiley face}
\usage{
smile(
  n,
  x0 = 0,
  y0 = 0,
  main_radius = 1,
  eye_radius = 0.25,
  mouth_radius = 0.7,
  mouth_start = 0.55,
  mouth_end = 0.95
)
}
\arguments{
\item{n}{the size of the sample}

\item{x0}{x coordinate of the center of the head, default to \code{0}}

\item{y0}{y coordinate of the center of the head, default to \code{0}}

\item{main_radius}{the radius of the head, default to \code{1}}

\item{eye_radius}{the radius of the eye, default to \code{0.25}}

\item{mouth_radius}{the radius of the mouth, default to \code{0.7}}

\item{mouth_start}{the radius of the mouth, default to \code{0.55}}

\item{mouth_end}{the radius of the mouth, default to \code{0.95}}
}
\value{
a list with two vectors of x and y of the samples coordinates.
The two list members are also named x and y and the list
itself is of class type \code{2dJointSample}.
}
\description{
\code{smile} simulates points in the plan
as if they were distributed uniformly on
a drawing of a smiling face.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\examples{
smile(5000)

}
\seealso{
\code{\link{frown}},\code{\link{triangle}},
\code{\link{circle}}, \code{\link{pizza}}, \code{\link{square}}.

Other Two-dimensional sampling functions.: 
\code{\link{Degenerate}()},
\code{\link{Exponential}()},
\code{\link{InversePareto}()},
\code{\link{Pareto}()},
\code{\link{frown}()},
\code{\link{plot.jointSample}()},
\code{\link{print.jointSample}()}
}
\concept{Two-dimensional sampling functions.}
