% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{triangle}
\alias{triangle}
\alias{triangle1}
\alias{triangle2}
\alias{triangle3}
\title{Simulate a Uniform distribution on a triangle}
\usage{
triangle(n, x0 = 0, y0 = 0)
}
\arguments{
\item{n}{the size of the sample}

\item{x0}{the size of the sample}

\item{y0}{the size of the sample}
}
\value{
A list with two vectors of x and y of the samples coordinates.
The two list members are also named x and y and the list
itself is of class type \code{2dJointSample}.
}
\description{
\code{triangle} simulates points in the plan as if they
were distributed uniformly on the inside of a triangle.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\examples{
mySample <- triangle(5000)

}
\seealso{
\url{https://www.r-project.org}

\code{\link{frown}} for products,
\code{\link{cumsum}} for cumulative sums, and
\code{\link{colSums}}/\code{\link{rowSums}}
marginal sums over high-dimensional arrays.

Other Two-dimensionalsampling functions.: 
\code{\link{circle}()},
\code{\link{pizza}()},
\code{\link{square}()}
}
\concept{Two-dimensionalsampling functions.}
