% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Uniform}
\alias{Uniform}
\alias{Uniform1}
\alias{Uniform2}
\alias{Uniform3}
\title{Uniform Distribution defaults to \code{[0, 1]}}
\usage{
Uniform(params)
}
\arguments{
\item{params}{the size of the sample}
}
\value{
a list with two vectors of x and y of the samples coordinates.
The two list members are also named x and y and the list
itself is of class type \code{2dJointSample}.
}
\description{
\code{Uniform} simulates points in the plan
as if they were distributed uniformly on
a drawing of a frowning face.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\examples{
Uniform(c(0, 1))

}
\seealso{
\code{\link{Degenerate}},\code{\link{Pareto}},
\code{\link{Exponential}}, \code{\link{InversePareto}}.
}
\concept{One-dimensional loss distributions.}
