% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{circle}
\alias{circle}
\alias{circle1}
\alias{circle2}
\alias{circle3}
\title{Simulate a Uniform distribution on a circle}
\usage{
circle(n, r = 1, x0 = 0, y0 = 0, from = 0, to = 1)
}
\arguments{
\item{n}{the size of the sample}

\item{r}{the radius of the circle, default to \code{1}}

\item{x0}{x coordinate of the center of the circle, default to \code{0}}

\item{y0}{y coordinate of the center of the circle, default to \code{0}}

\item{from}{starting point of the arc as a fraction of the whole
circumference starting from the rightmost point and
drawing the circle in a counter-clockwise direction,
default to \code{0}}

\item{to}{ending point of the arc as a fraction of the whole
circumference starting from the rightmost point and
drawing the circle in a counter-clockwise direction,
default to \code{0}}
}
\value{
A list with two vectors of x and y of the samples coordinates.
The two list members are also named x and y and the list
itself is of class type \code{2dJointSample}.
}
\description{
\code{circle} simulates points in the plan as if they
were distributed uniformly on a circle arc.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\examples{
circle(5000)

}
\seealso{
\code{\link{frown}},\code{\link{smile}},
\code{\link{triangle}}, \code{\link{pizza}}, \code{\link{square}}.

Other Two-dimensionalsampling functions.: 
\code{\link{pizza}()},
\code{\link{square}()},
\code{\link{triangle}()}
}
\concept{Two-dimensionalsampling functions.}
