% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{copula}
\alias{copula}
\title{copula}
\usage{
copula(joint_sample)
}
\arguments{
\item{jointSample}{An object of class "jointSample" such as
created by a call to smile or frown}
}
\value{
The empirical copula model of thejont sample
input \deqn{\int_0^1xdx= \frac{x^2}{2} + C}
and
\deqn{p(x) = \frac{\lambda^x e^{-\lambda}}{x!}}{%p(x) =
\lambda^x exp(-\lambda)/x!} for \eqn{x = 0, 1, 2, \ldots}.
}
\description{
\code{copula} generates the empirical copula
corresponding to the joint sample passed in
as input
}
\examples{
copula(joint_sample)
}
