% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{copula}
\alias{copula}
\title{The \code{copula} function}
\usage{
copula(joint_sample)
}
\arguments{
\item{joint_sample}{An object of class \code{jointSample} such as
created by a call to \code{smile} or \code{frown}}
}
\value{
The empirical copula model of the joint sample
}
\description{
\code{copula} generates the empirical copula
corresponding to the joint sample passed in
as input
}
\examples{
sum(9,5)

}
\seealso{
\url{https://www.r-project.org}
}
