% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVX.R
\docType{package}
\name{CVX}
\alias{CVX}
\title{CVX: A package for Copula Visual eXploration}
\description{
The CVX package provides three categories of functions:
\itemize{
\item functions to easily generate 2-dimensionaljoint samples
with markedly different dependency structures
\item functions to generate copulas from joint samples
as well as joint samples from copulas
\item functions to plot and toanimate copulas or joint samples
}
}
\section{Simple two dimensional joint sample models}{

\itemize{
\item \code{\link{frown}}
\item \link{smile}
\item \link{circle}
\item \link{pizza}
\item \link{square}
\item \link{triangle}
}
}

\section{Copula models}{

\itemize{
\item \code{CommonShock}
\item \code{ErlangMixtures}
}
}

\section{Utility functions}{

\itemize{
\item \code{PlotSample}
\item \code{CopMovie}
}
}

\section{Other functions}{

\itemize{
\item \code{OtherFunction1}
\item \code{OtherFunction2}
}
}

