% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{Gaussian}
\alias{Gaussian}
\alias{Gaussian1}
\alias{Gaussian2}
\alias{Gaussian3}
\title{Simulate a joint normal distribution}
\usage{
Gaussian(n, mu, sigmasq, rho)
}
\arguments{
\item{n}{the size of the sample}

\item{mu}{the vector of means}

\item{sigmasq}{the covariance matrix}

\item{rho}{correlation coefficient}
}
\value{
A list with two vectors of x and y of the samples
coordinates. The two list members are also named
x and y and the list itself is of class type
\code{2dJointSample}.
}
\description{
\code{Gaussian} simulates points in the plan as if they
were distributed as joint Normal random variables.
}
