% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{my_cdf}
\alias{my_cdf}
\alias{my_cdf1}
\alias{my_cdf2}
\alias{my_cdf3}
\title{Generate the empirical copula from a joint random sample}
\usage{
my_cdf(x)
}
\arguments{
\item{x}{an object of type \code{2DSample}}
}
\value{
The empirical copula model of the joint sample
as a list with two vectors of x and y of the
samples coordinates.
}
\description{
\code{my_cdf} generates the empirical copula
corresponding to the joint sample passed in
as input
}
\examples{
my_cdf(frown(1000)$x)

}
