% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{square}
\alias{square}
\alias{square1}
\alias{square2}
\alias{square3}
\title{Draws points uniformly distributed on the perimeter
of a square}
\usage{
square(n)

square(n)
}
\arguments{
\item{n}{the size of the sample}
}
\value{
A list with two vectors of x and y of the
samples coordinates. The two list members are
also named x and y and the list itself is of
class type \code{2DSample}.

A list with two vectors of x and y of the
samples coordinates. The two list members are
also named x and y and the list itself is of
class type \code{2DSample}.
}
\description{
\code{square} simulates points in the plan as if they
were distributed uniformly on the perimeter of a square.

\code{square} simulates points in the plan as if they
were distributed uniformly on the perimeter of a square.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.

This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\examples{
square(5000)

square(5000)

}
\seealso{
Other Two-dimensional sampling functions: 
\code{\link{circle}()},
\code{\link{frown}()},
\code{\link{pizza}()},
\code{\link{smile}()},
\code{\link{triangle}()}

Other Two-dimensional sampling functions: 
\code{\link{circle}()},
\code{\link{frown}()},
\code{\link{pizza}()},
\code{\link{smile}()},
\code{\link{triangle}()}
}
\concept{Two-dimensional sampling functions}
