% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_samples.R
\name{triangle}
\alias{triangle}
\alias{triangle1}
\alias{triangle2}
\alias{triangle3}
\title{Draws points uniformly distributed inside a triangle}
\usage{
triangle(n)
}
\arguments{
\item{n}{the size of the sample}
}
\value{
A list with two vectors of x and y of the
samples coordinates. The two list members are
also named x and y and the list itself is of
class type \code{2DSample}.
}
\description{
\code{triangle} simulates points in the plan as if they
were distributed uniformly on the inside of a triangle.
}
\details{
This and the list of functions below allows the
user to quickly generate two-dimensional samples
with very different dependency structures.
}
\examples{
triangle(5000)

}
\seealso{
Other Two-dimensional sampling functions: 
\code{\link{circle}()},
\code{\link{frown}()},
\code{\link{pizza}()},
\code{\link{smile}()},
\code{\link{square}()}
}
\concept{Two-dimensional sampling functions}
